Session_12/Spaces/requirements.txt
torch
albumentations
pytorch-lightning
torchvision
torch-lr-finder
grad-cam
pillow
numpy
gradio